src/flowerpredict/requirements.txt
azure-identity
azure-storage-blob
fastapi
pydantic
uvicorn
joblib
python-dotenv
Pillow
numpy
scikit-learn
tensorflow
python-multipart
